move `build_changes_for_partition` to `struct OverPartition`
